+++ example/webpack/package.json
@@ -14,6 +14,7 @@
     "@babel/preset-env": "^7.21.4",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.21.4",
+    "@kuma-ui/webpack-plugin": "workspace:^",
     "@types/react-dom": "^18.2.1",
     "babel-loader": "^9.1.2",
     "css-loader": "^6.7.3",
@@ -26,7 +27,6 @@
   },
   "dependencies": {
     "@kuma-ui/core": "workspace:^",
-    "@kuma-ui/webpack-loader": "workspace:^",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "scheduler": "^0.23.0"
