--- packages/webpack-loader/src/index.ts
@@ -1,4 +1,0 @@
-// import KumaUIWebpackPlugin from "./plugin";
-import kumaUiLoader from "./loader";
-// export default KumaUIWebpackPlugin;
-export default kumaUiLoader;
