+++ packages/webpack-plugin/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@kuma-ui/webpack-loader",
+  "name": "@kuma-ui/webpack-plugin",
   "version": "0.0.2",
   "description": "🐻 Kuma UI is a utility-first, zero-runtime CSS-in-JS library that offers an outstanding developer experience and optimized performance.",
   "repository": {
