+++ packages/webpack-plugin/src/index.ts
@@ -1,0 +1,2 @@
+import { default as KumaUIWebpackPlugin } from "./plugin";
+export default KumaUIWebpackPlugin;
